Put back everything into the wsgi.py file.
